Add files via upload
Machine Learning helps in predicting the Heart diseases, and the predictions made are quite accurate.The project involved analysis of the heart disease patient dataset with proper data processing. 
Then, different models were trained and and predictions are made with different algorithms KNN, Decision Tree, Random Forest,SVM and Logistic Regression. This is the jupyter notebook code.
I've used a variety of Machine Learning algorithms, implemented in Python, to predict the presence of heart disease in a patient. This is a classification problem, with input features as a variety of parameters, and the target variable as a binary variable, predicting whether heart disease is present or not.

Machine Learning algorithms used:

Logistic Regression (Scikit-learn)
Support Vector Machine (Linear) (Scikit-learn)
K-Nearest Neighbours (Scikit-learn)
Decision Tree (Scikit-learn)
Random Forest (Scikit-learn)